消除效果

+++ PopStar Sprite/ThePopSprite.h
@@ -1,9 +1,8 @@
 #include "cocos2d.h"
 USING_NS_CC;
 class PopSprite : public Sprite{
-	//初始化游戏的数值，宽度，高度，x轴，y轴
-	
 public:
+	//初始化游戏的数值，宽度，高度，x轴，y轴
 	static PopSprite* createPopSprite(int color,int width,int height,float popSpriteX,float popSpriteY);
 	virtual bool init();
 	//获取颜色的函数
@@ -26,7 +25,7 @@ private:
 	//颜色
 	int iColor;
 	//私有化的初始方法
-	bool privateInit(int color,int width, int height, int popSpriteX, int popSpriteY);
+	bool privateInit(int color,int width, int height, float popSpriteX, float popSpriteY);
 	//显示的颜色
 	LayerColor* ColorBlock;
 };
